pyproject.toml
[project]
name = "expense-tracker-api"
version = "0.1.0"
description = "A Flask-based REST API for expense tracking"
requires-python = ">=3.10"
dependencies = [
    "flask==3.0.3",
    "flask-sqlalchemy==3.1.1",
    "flask-migrate==4.0.7",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.0.1",
    "flask-socketio==5.5.1",
    "python-socketio==5.12.1",
    "pydantic==2.10.5",
    "werkzeug==3.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"